backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
asyncpg==0.29.0
httpx==0.25.2
coverage==7.4.0
Authlib==1.3.0
email-validator==2.1.1
itsdangerous==2.1.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
